// seatunnel-engine/seatunnel-engine-ui/src/service/trace/types.ts
/*
 * Licensed to the Apache Software Foundation (ASF) under one or more
 * contributor license agreements.  See the NOTICE file distributed with
 * this work for additional information regarding copyright ownership.
 * The ASF licenses this file to You under the Apache License, Version 2.0
 * (the "License"); you may not use this file except in compliance with
 * the License.  You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface TraceSummary {
  traceId: number
  sinkTaskId: number
  jobId: string
  tableId: string
  createdTimeMs: number
  receivedTimeMs: number
  startTsMs: number
  entryCount: number
}

export interface TraceEntry {
  index: number
  stage: number
  taskId: number
  tsMs: number
  workerAddress?: string
  taskGroupName?: string
  taskClass?: string
}

export interface TraceDetail {
  trace: TraceSummary
  entries: TraceEntry[]
}